Update definitions
* separate files
* remove unnecessary `{[key: string]: any}`
* from singular to plural
